DTO/Service/Controller/Repository Long→Integer 전수 수정
DB 스키마 기준 int 타입 PK/FK (category_id, company_id, region_id, parent_id)를
참조하는 모든 DTO, Service 인터페이스/구현체, Controller, Repository에서
Long → Integer 타입으로 일괄 수정.

수정 범위: DTO 10개, Service 인터페이스 7개, Service 구현체 11개,
           Controller 12개, Repository 1개

+++ src/main/java/com/honeyrest/honeyrest_host/controllerAdmin/AccommodationController.java
@@ -353,7 +353,7 @@ public class AccommodationController {
     // 지역(대)의 하위(소) 목록 JSON (select 연동)
     @GetMapping(value = "/regions/children", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public List<Map<String, Object>> regionChildren(@RequestParam("parentId") Long parentId) {
+    public List<Map<String, Object>> regionChildren(@RequestParam("parentId") Integer parentId) {
         return regionRepository.findByParentId(parentId).stream()
                 .map(r -> Map.<String, Object>of(
                         "id", r.getRegionId(),
