+++ src/main/java/com/honeyrest/honeyrest_host/controllerOwner/AccommodationController.java
@@ -26,7 +26,7 @@ public class AccommodationController {
     private final ORoomService roomService;
 
     @GetMapping({"/accommodation/list", "/company/{companyId}/accommodations"})
-    public String accommodations(@PathVariable(required = false) Long companyId,
+    public String accommodations(@PathVariable(required = false) Integer companyId,
                                  @ModelAttribute PageRequestDTO pageRequestDTO,
                                  Model model) {
 
@@ -48,7 +48,7 @@ public class AccommodationController {
 
     @GetMapping("/accommodation/inActive/list")
     public String accommodationsInActiveList(@ModelAttribute PageRequestDTO pageRequestDTO,
-                                             @RequestParam(required = false) Long companyId,
+                                             @RequestParam(required = false) Integer companyId,
                                              Model model) {
 
         PageResponseDTO<AccommodationDTO> responseDTO =
@@ -64,7 +64,7 @@ public class AccommodationController {
 
 
     @GetMapping("/accommodation/create")
-    public String createAccommodation(@RequestParam Long companyId, Model model) {
+    public String createAccommodation(@RequestParam Integer companyId, Model model) {
         model.addAttribute("companyId", companyId);
         model.addAttribute("companies", companyService.getAllCompanies());
         model.addAttribute("categories", accommodationCategory.getAllAccommodationCategory());
@@ -128,14 +128,14 @@ public class AccommodationController {
     // 회사 검색 API (자동완성용)
     @GetMapping("/accommodation/companyId/search")
     @ResponseBody
-    public List<AccommodationDTO> searchAccommodation(@RequestParam Long companyId, @RequestParam String keyword) {
+    public List<AccommodationDTO> searchAccommodation(@RequestParam Integer companyId, @RequestParam String keyword) {
         return accommodationService.searchByNameContaining(companyId, keyword).stream().filter(a-> a.getStatus().equalsIgnoreCase("active")).toList();
     }
 
     // 회사 검색 API (자동완성용)
     @GetMapping("/accommodation/inActive/search")
     @ResponseBody
-    public List<AccommodationDTO> searchInActiveAccommodation(@RequestParam Long companyId, @RequestParam String keyword) {
+    public List<AccommodationDTO> searchInActiveAccommodation(@RequestParam Integer companyId, @RequestParam String keyword) {
         return accommodationService.searchByNameContaining(companyId, keyword).stream().filter(a-> !a.getStatus().equalsIgnoreCase("active")).toList();
     }
