+++ src/main/java/com/honeyrest/honeyrest_host/serviceAdmin/CompanyService.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public interface CompanyService {
     CompanyDTO create(CompanyDTO dto);
-    CompanyDTO getById(Long id);
+    CompanyDTO getById(Integer id);
     List<CompanyDTO> getAll();
-    CompanyDTO update(Long id, CompanyDTO dto);
-    void delete(Long id);
+    CompanyDTO update(Integer id, CompanyDTO dto);
+    void delete(Integer id);
 
     CompanyDTO getByUserEmail(String email);
