+++ src/main/java/com/honeyrest/honeyrest_host/serviceAdmin/RegionServiceImpl.java
@@ -27,13 +27,13 @@ public class RegionServiceImpl implements RegionService {
     }
 
     @Override
-    public List<RegionDTO> listSubRegions(Long parentId) {
+    public List<RegionDTO> listSubRegions(Integer parentId) {
         List<Region> regions = regionRepository.findByParentId(parentId);
         return regions.stream().map(RegionDTO::of).toList();
     }
 
     @Override
-    public RegionDTO get(Long regionId) {
+    public RegionDTO get(Integer regionId) {
         Region region = regionRepository.findById(regionId)
                 .orElse(null); // 못 찾으면 null 리턴(혹은 예외 던져도 OK)
         return (region != null ? RegionDTO.of(region) : null);
